Further break down e2e logic (#703)
Fixes #691.

- Refactors e2e into smaller steps. 
- Introduces a `run_partial` method which takes a `prefix` argument,
indicating when to stop the run of the pipeline. According to this
argument, different parts of the state are yielded back. This part isn't
too pretty type-wise, but ultimately I think it's not dangerous and a
reasonable compromise for the moment.

Later edit: I've also included a benchmark for witness generation of the
Fibonacci program, since it is a consumer of this refactor and helps to
exemplify it.