Suggestions for [Feat] Calculate proof size and number of hashes (#771)
Here are some suggestions for
https://github.com/scroll-tech/ceno/pull/712

This builds on https://github.com/scroll-tech/ceno/pull/770 and the diff
will get smaller, once #770 is in master.

The main idea here is to remove the need to have a reference count and
also simultaneously remove the need for a manual drop. We enlist the
borrow checker's help to make sure that we only show the statistics,
once everything has been added up. (No more `sync_stat` necessary,
either.)