Revise IsLtConfig instance assignment interface
Existing interface allowed assignment by u64 and by SWord, which is an alias for
i32.  In this commit, the SWord assignment is changed to i64 assignment, which
enables use in more general contexts (i32 values can always be cast to i64
without changing the results of comparison, which is what the underlying
datatype is used for in the assignment), and both signed and unsigned
assignment is derived from a single assignment function taking field elements
and an explicit `is_lt` boolean input.

The i64 interface in particular will be useful for assignments needed by the
signed DIV and REM opcodes, which have edge cases whose values lie outside of
the i32 data range.