add whir as a crate in ceno via git subtree (#841)
re-open of #840 due to the base branch change
 
integrate whir as git subtree.

### why subtree
- This enable us to plug `whir` as a crate in ceno, thus it can share
most of libraries with ceno
- review everything just in one ceno PR
- retain approach sync back to upstream via command `git subtree push
--prefix=whir https://github.com/scroll-tech/whir <feature branch>`. The
timing to do that is decouple with our development in Ceno. We only need
that when going to sync back to upstream

With this PR, we can start to do refactor on whir, e.g. 
- migrate field to plonky3
- unify transcript
- share MLE + sumcheck from Ceno
- ...

---------

Co-authored-by: Yuncong Zhang <yczhangsjtu@gmail.com>