bn254_syscalls  (#828)
Adds interface for precompiles pertaining to the `bn254` elliptic curve.
Mostly follows the structure of the existing precompiles.

Caveats:
- This PR changes the mock `ecall` circuit to accept arbitrary memory
addresses, as opposed to consecutive addresses as it does now. As
discussed with @naure, this requires [more
attention](https://github.com/Inversed-Tech/ceno/issues/5); but I think
the review can happen in parallel
- The guest examples use debug messages for testing purposes. Debug
messages are currently incompatible with proving. I'm trying to make the
logging conditional via a flag in the input, but I need to debug some
weird behavior.