Add rotation verifier (#965)
Here is a summary of the updates in this PR:
1. Correct the previous typo in the rotation prover, and add a rotation
verifier.
2. Change the selector formular:
0(rt) = \sum_{b \in \{0, 1\}^n} sel(rt, b) * (\sum_i alpha^i *
rotated[i](b) - target[i](b)))
   where sel(rt, b) = eq(rt[5..], b[5..]) * sel'(rt[..5], b[..5])
   where sel'(rt', b') = b' in rotated[0..24] ? eq(rt', b') : 0.
3. Refine the default method `gkr_witness`. Previously, it assumes the
`phase1_witnesses` are connected to the input layer in the same order.
In a more general case, the `phase1` witnesses can be connected to the
wires in any layer, which is set by `allocate_opening` method. I modify
`gkr_witness` in consider of this case.

| Benchmark                  | My PC Running Time | Median Change (%) |
|----------------------------|--------------------|-------------------|
| prove_keccak_lookup_f_4096 | 4.7770 s           | -10.773%          |
| prove_keccak_lookup_f_8192 | 11.178 s           | -1.9963%          |

---------

Co-authored-by: kunxian xia <xiakunxian130@gmail.com>