Refactor GKR IOP Circuit Construction (#970)
This PR includes the following changes:
1. Support shorter fixed polys (see comments in the code).
2. Simplify the input layer witness generation, assuming the committed
Phase 1 witnesses only occur in the input layer, and the evaluations to
be opened are corresponding to the input layer evaluation expressions.
- Note: It doesn't support the multi-layer logup circuit because it
requires merging the left and right evaluations before opening a
polynomial.
4. Refactor `LayerConstraintSystem` to construct GKR layers. It will pad
fixed polys after wit polys.
5. Rewrite Keccak circuit (both lookup and bitwise version).

### Future work
- (from Ming) fixed poly `RC` need to update from verifier side. In this
case, better version of `RC` is structural_witin

---------

Co-authored-by: dreamATD <tianyi.liu.08@gmail.com>
Co-authored-by: sm.wu <hero78119@gmail.com>