use par_iter_mut instead of scoped spawns in phase1 of sumcheck (#1011)
Hi all, I have encountered a racy behaviour of workers in phase1 of
sumcheck prover that essentially forbids running two sumcheck proving
processes in a parallel because of a potential deadlock. The nature of
the deadlock is that the rayon tasks are not that well designed for
communicating between each other using channels, as discussed
[here.](https://github.com/rayon-rs/rayon/issues/1258).

`ceno` starts racing if two or more provers are being run in the
following situation:
* consider you have a machine such as a 12-core Mac
* you want to launch two proving processes each with 8 workers in phase1
* rayon pool by default will likely contain 12 worker threads
* first ceno prover manages to spawn 7 worker tasks
* second ceno worker manages to spawn another 5 tasks
* however the main worker in both of them waits for evaluations of the
first iteration from _all_ workers and _all_ workers do not exist in
eather of the provers.
* both processes are now hanging deadlocked.

A repro can be found
[here](https://github.com/montekki/ceno-sum-rayon-test).

This patch uses `par_iter_mut` instead of the existing workload
architecture and does not use blocking channels. It also simplifies code
somewhat unifying it for main and non-main workers and removing the
syncronised transcript structure.