assert lt msb limb in strict range check (#1018)
### background
this issue was found during development of PR
https://github.com/scroll-tech/ceno/pull/1010
In current master branch, `IsLtGadget` assume the |lhs - rhs| difference
are in multiple of 16 bits, see here


https://github.com/scroll-tech/ceno/blob/4b4de4f670ec966b6d51f938bacde4a314471983/gkr_iop/src/gadgets/is_lt.rs#L217

This assumption will bring some potiential soundness as the condition
are much relaxation version.
To fix this, the strict condition is just need the `log_2(|lhs - rhs|)`
bits to constrain the `diff` range, which means we just improve
granularity range check with just `log_2(|lhs - rhs|) % 16` on
most-significant-limb instead of full u16 check.

This PR also related to future work
https://github.com/scroll-tech/ceno/issues/1019 so we could have more
flexible range check on a single table.

### other side note
Migrate from u16 multiple to `max_bits` make it work when testing on
smaller field, e.g. babybear. An known issue in
https://github.com/scroll-tech/ceno/pull/1010 unittest is
`ts`(timestamp) in offline memory check are assume to be u32, thus, max
`diff` could be `1<<32`, however `BabyBear::from_u32(1<<32)` doesnt work
because babybear can only encode number < `2^31 - 2^27 + 1 `. So in this
PR, we add a workaround of hardcap on max timestamp to < 2^30. As each
cycle the incr of timestamp could be up to 4 = 2^2, so we can only deal
with maximum of (30 - 2) => 2^28 cycles which is roughly 256M cycles. To
increase maximal cycles support, see issue
https://github.com/scroll-tech/ceno/issues/1000
 
### babybear
in this PR we also add more babybear opcode test.

### circuits stats
From circuit stats, before/after change, all circuit constrains remain
the same, thus we skip benchmark for it